bugfix for accidental circular dependancy post-merge
