+++ roundabout/search/tables.py
@@ -32,7 +32,6 @@ from roundabout.calibrations.models import CalibrationEvent
 from roundabout.configs_constants.models import ConfigEvent
 from roundabout.inventory.models import Inventory, Action
 from roundabout.parts.models import Part
-from roundabout.search.user_search import ActionUserTable
 
 
 def trunc_render(length=100, safe=False, showable=True, targets=None, bold_target=True):
@@ -207,11 +206,51 @@ class ActionTable(SearchTable):
         model = Action
         fields = ['object_type', 'object', 'action_type', 'user__name', 'created_at', 'detail']
         base_shown_cols = fields
+        
     user__name = Column(verbose_name='User')
     object = Column(verbose_name='Associated Object', accessor='object_type')
-    render_object = ActionUserTable.render_object  # yay for weird ways of implementing code re-use
+
     def render_user__name(self,record):
         return record.user.name or record.user.username
+
+    def render_object(self,record):
+        html_string = '<a href={url}>{text}</a>'
+        parent_obj = record.get_parent()
+        if isinstance(parent_obj,(CalibrationEvent,ConfigEvent,ConstDefaultEvent)):
+            parent_obj = parent_obj.inventory
+        elif isinstance(parent_obj,(ConfigNameEvent,CoefficientNameEvent)):
+            parent_obj = parent_obj.part
+        elif isinstance(parent_obj,ConfigDefaultEvent):
+            parent_obj = parent_obj.assembly_part
+
+        if isinstance(parent_obj, Inventory):
+            inv_url = reverse("inventory:inventory_detail", args=[parent_obj.pk])
+            html_string = html_string.format(url=inv_url, text=parent_obj)
+            return format_html(html_string)
+        elif isinstance(parent_obj,Build):
+            build_url = reverse("builds:builds_detail", args=[parent_obj.pk])
+            html_string = html_string.format(url=build_url, text=parent_obj)
+            return format_html(html_string)
+        elif isinstance(parent_obj,Deployment):
+            build_url = reverse("builds:builds_detail", args=[record.deployment.build.pk])
+            deployment_anchor = '#deployment-{}-'.format(record.deployment.pk)  # doesn't work, anchor doesn't exist
+            deployment_anchor = '#deployments'  # next best anchor that does work
+            html_string = html_string.format(url=build_url+deployment_anchor, text=parent_obj)
+            return format_html(html_string)
+        elif isinstance(parent_obj,InventoryDeployment):
+            inv_url = reverse("inventory:inventory_detail", args=[parent_obj.inventory.pk])
+            html_string = html_string.format(url=inv_url, text=parent_obj)
+            return format_html(html_string)
+        elif isinstance(parent_obj,Part):
+            build_url = reverse("parts:parts_detail", args=[parent_obj.pk])
+            html_string = html_string.format(url=build_url, text=parent_obj)
+            return format_html(html_string)
+        elif isinstance(parent_obj,AssemblyPart):
+            assy_url = reverse("assemblies:assemblypart_detail", args=[parent_obj.pk])
+            html_string = html_string.format(url=assy_url, text=parent_obj)
+            return format_html(html_string)
+        else:
+            return ''
 
 
 class CalibrationTable(SearchTable):
